refactor: clean up editor transition code
- Remove unused newComponentLayers param from updateLayersWithComponent
- Drop redundant containsComponent pre-check in updateComponentOnLayers
- Extract scheduleIdle to shared lib/schedule-idle.ts utility
- Deduplicate parentComponentId resolution in LayerRenderer
- Document inflightDraftLoads reset requirement

Made-with: Cursor
