+++ stores/useComponentsStore.ts
@@ -13,26 +13,9 @@ import {
   cleanLayersForComponentCreation,
 } from '@/lib/layer-utils';
 import { detachStyleFromLayers, updateLayersWithStyle } from '@/lib/layer-style-utils';
+import { scheduleIdle } from '@/lib/schedule-idle';
 import { generateId } from '@/lib/utils';
 import type { Component, Layer } from '@/types';
-
-/**
- * Schedule a callback to run when the browser is idle, so it does not block
- * the current interaction. Falls back to setTimeout in environments without
- * requestIdleCallback (Safari, SSR) and to a sync call on the server.
- */
-function scheduleIdle(callback: () => void): void {
-  if (typeof window === 'undefined') {
-    callback();
-    return;
-  }
-  const ric = (window as unknown as { requestIdleCallback?: (cb: () => void, opts?: { timeout: number }) => number }).requestIdleCallback;
-  if (typeof ric === 'function') {
-    ric(callback, { timeout: 1000 });
-  } else {
-    setTimeout(callback, 0);
-  }
-}
 
 /** Remove variableLinks entries that point TO a given variable ID (as parent target). */
 function removeVariableLinksPointingTo(layer: Layer, targetVariableId: string): Layer {
